feat(*): fixes after review
