+++ src/client/types.ts
@@ -1,13 +1,6 @@
-export type NativeParams = {
-    appVersion: string;
-    title?: string;
-    // В ранних версиях iOS приложение не пробрасывет схему приложения в URL в прод окружении.
-    // Для таких версий есть мэппинг `./constants` → `versionToIosAppId`.
-    iosAppId?: string;
-    theme: string;
-    nextPageId: number | null;
-    originalWebviewParams: string;
-};
+import { NativeParamsType } from "../shared/types";
+
+export type NativeParams = NativeParamsType;
 
 export type NativeFeatureKey =
     // Возможность работы с геолокацией.
