+++ src/index.ts
@@ -1,2 +1,2 @@
-export { BridgeToNative } from './client/bridge-to-native';
-export { NativeParams, Theme, Environment, NativeFeatureKey, PdfType } from './client/types';
+export * from './client'
+export * from './server';
