+++ src/server/constants.ts
@@ -1,13 +1,7 @@
 export const THEME_QUERY = 'theme';
 export const TITLE = 'title';
-
 export const WEBVIEW_IOS_APP_ID_QUERY = 'applicationId';
 export const WEBVIEW_IOS_APP_VERSION_QUERY = 'device_app_version';
 export const WEBVIEW_WITHOUT_LAYOUT_QUERY = 'without_layout';
 export const WEBVIEW_NEXT_PAGE_ID_QUERY = 'nextPageId';
-
-export const CAPSULE_UA_SUBSTR = 'Capsule';
-export const AKEY_UA_SUBSTR = 'AKEY';
-export const VOSKHOD_UA_SUBSTR = 'VOSKHOD';
-
 export const NATIVE_PARAMS_COOKIE_NAME = 'app_native_params';
