+++ src/server/is-webview-environment.ts
@@ -1,8 +1,7 @@
 import {
     extractAppVersion,
     extractUserAgent,
-    isAkeyWebview,
-    extractNativeParamsFromCookies
+    extractNativeParamsFromCookieHeader
 } from './utils';
 import { RequestHeaderType } from './types';
 import { versionPattern, webviewUaIOSPattern } from './reg-exp-patterns';
@@ -11,10 +10,6 @@ export const isWebviewByUserAgent = (
     userAgent: string,
     appVersion: string | undefined,
 ) => {
-    if (userAgent && isAkeyWebview(userAgent)) {
-        return false;
-    }
-
     return (
         (appVersion && versionPattern.test(appVersion)) || !!userAgent?.match(webviewUaIOSPattern)
     );
@@ -23,14 +18,14 @@ export const isWebviewByUserAgent = (
 export const isWebviewByCookies = (nativeParamsFromCookies: Record<string, any> | null) => {
     return !!(nativeParamsFromCookies && nativeParamsFromCookies.isWebview)
 }
-export const checkIsWebview = (
+export const isWebviewEnvironment = (
     request: RequestHeaderType,
 ): boolean => {
     const userAgent = extractUserAgent(request);
 
     // `app-version` в заголовках – индикатор вебвью. В iOS есть только в первом запросе от webview
     const appVersion = extractAppVersion(request);
-    const nativeParamsFromCookies = extractNativeParamsFromCookies(request);
+    const nativeParams = extractNativeParamsFromCookieHeader(request);
 
-    return isWebviewByCookies(nativeParamsFromCookies) || isWebviewByUserAgent(userAgent, appVersion);
+    return isWebviewByCookies(nativeParams) || isWebviewByUserAgent(userAgent, appVersion);
 };
