+++ src/server/types.ts
@@ -1,16 +1,11 @@
+import type { NativeParamsType } from '../shared/types';
 export type RequestHeaderType = Record<string, any>;
 
 export type EmptyNativeParams = {
     isWebview: false;
 };
 
-export type NativeParams = {
-    appVersion: string;
-    iosAppId?: string;
+export type NativeParams = NativeParamsType & {
     isWebview: true;
-    theme: 'dark' | 'light';
-    title: string;
     withoutLayout: boolean;
-    originalWebviewParams: string;
-    nextPageId: number | null;
 };
